archiver/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Haoming Xia <mrmist3r311@gmail.com>"]
packages = [
    { include = "db" },
    { include = "crawler" },
    { include = "api" },
    { include = "modules" }
]

[tool.poetry.dependencies]
python = "^3.10"
google-api-python-client = "^2.54.0"
SQLAlchemy = "^1.4.39"
requests = "^2.28.1"
Flask = "^2.1.3"
python-dotenv = "^0.20.0"
pydantic = "^1.9.1"
sqlalchemy2-stubs = "^0.0.2-alpha.24"
google-api-python-client-stubs = "^1.11.0"
python-dateutil = "^2.8.2"
xmltodict = "^0.13.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pylint = "^2.14.5"
black = { version = "^22.6.0", allow-prereleases = true }

[tool.poetry.scripts]
seed = "db.seed:main"
update_video_status = "crawler.tasks.video_status:video_status"
update_videolist_nijisanji = "crawler.tasks.videolist_nijisanji:videolist_nijisanji"
update_videolist_rss = "crawler.tasks.videolist_rss:videolist_rss"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
